+++ libs/compiler/codes.c
@@ -1481,9 +1481,9 @@ void tables_and_codes(const syntax *const sx, const char *const path)
 	}
 
 	uni_printf(&io, "\n\n%s\n", "functions");
-	for (size_t i = 1; (int)i <= sx->funcnum; i++)
+	for (size_t i = 1; i <= sx->funcnum; i++)
 	{
-		uni_printf(&io, "fun %zi) %i\n", i, sx->functions[i]);
+		uni_printf(&io, "fun %zi) %zi\n", i, sx->functions[i]);
 	}
 
 	uni_printf(&io, "\n%s\n", "iniprocs");
